ExceptionConverter alignment
