+++ src/Cuemon.Xml/Extensions/Serialization/Converters/XmlConverterDecoratorExtensions.cs
@@ -9,7 +9,7 @@ using Cuemon.Xml.Linq;
 
 namespace Cuemon.Xml.Serialization.Converters
 {
-/// <summary>
+    /// <summary>
     /// Extension methods for the <see cref="XmlConverter"/> class tailored to adhere the decorator pattern.
     /// </summary>
     /// <seealso cref="IDecorator{T}"/>
@@ -176,7 +176,7 @@ namespace Cuemon.Xml.Serialization.Converters
                 if (options.IncludeFailure)
                 {
                     writer.WriteStartElement("Failure");
-                    WriteException(writer, descriptor.Failure, options.IncludeStackTrace);
+                    new ExceptionConverter(options.IncludeStackTrace).WriteXml(writer, descriptor.Failure);
                     writer.WriteEndElement();
                 }
                 writer.WriteEndElement();
@@ -299,100 +299,16 @@ namespace Cuemon.Xml.Serialization.Converters
         /// Adds an <see cref="Exception" /> XML converter to the enclosed <see cref="T:IList{XmlConverter}"/> of the specified <paramref name="decorator"/>.
         /// </summary>
         /// <param name="decorator">The <see cref="T:IDecorator{IList{XmlConverter}}" /> to extend.</param>
-        /// <param name="includeStackTraceFactory">The function delegate that is invoked when it is needed to determine whether the stack of an exception is included in the converted result.</param>
+        /// <param name="includeStackTrace">The value that determine whether the stack of an exception is included in the converted result.</param>
         /// <returns>A reference to <paramref name="decorator"/> after the operation has completed.</returns>
         /// <exception cref="ArgumentNullException">
         /// <paramref name="decorator"/> cannot be null.
         /// </exception>
-        public static IDecorator<IList<XmlConverter>> AddExceptionConverter(this IDecorator<IList<XmlConverter>> decorator, Func<bool> includeStackTraceFactory)
+        public static IDecorator<IList<XmlConverter>> AddExceptionConverter(this IDecorator<IList<XmlConverter>> decorator, bool includeStackTrace)
         {
             Validator.ThrowIfNull(decorator, nameof(decorator));
-            decorator.AddXmlConverter<Exception>((writer, exception, _) =>
-            {
-                WriteException(writer, exception, includeStackTraceFactory?.Invoke() ?? false);
-            });
+            decorator.Inner.Add(new ExceptionConverter(includeStackTrace));
             return decorator;
-        }
-
-        private static void WriteException(XmlWriter writer, Exception exception, bool includeStackTrace)
-        {
-            var exceptionType = exception.GetType();
-            writer.WriteStartElement(Decorator.Enclose(exceptionType.Name).SanitizeXmlElementName());
-            if (exceptionType.Namespace != null) { writer.WriteAttributeString("namespace", exceptionType.Namespace); }
-            WriteExceptionCore(writer, exception, includeStackTrace);
-            writer.WriteEndElement();
-        }
-
-        private static void WriteExceptionCore(XmlWriter writer, Exception exception, bool includeStackTrace)
-        {
-            if (!string.IsNullOrEmpty(exception.Source))
-            {
-                writer.WriteElementString("Source", exception.Source);
-            }
-
-            if (!string.IsNullOrEmpty(exception.Message))
-            {
-                writer.WriteElementString("Message", exception.Message);
-            }
-
-            if (exception.StackTrace != null && includeStackTrace)
-            {
-                writer.WriteStartElement("Stack");
-                var lines = exception.StackTrace.Split(new[] { Alphanumeric.NewLine }, StringSplitOptions.RemoveEmptyEntries);
-                foreach (var line in lines)
-                {
-                    writer.WriteElementString("Frame", line.Trim());
-                }
-                writer.WriteEndElement();
-            }
-
-            if (exception.Data.Count > 0)
-            {
-                writer.WriteStartElement("Data");
-                foreach (DictionaryEntry entry in exception.Data)
-                {
-                    writer.WriteStartElement(Decorator.Enclose(entry.Key.ToString()).SanitizeXmlElementName());
-                    writer.WriteString(Decorator.Enclose(entry.Value.ToString()).SanitizeXmlElementText());
-                    writer.WriteEndElement();
-                }
-                writer.WriteEndElement();
-            }
-
-            var properties = Decorator.Enclose(exception.GetType()).GetRuntimePropertiesExceptOf<AggregateException>().Where(pi => !Decorator.Enclose(pi.PropertyType).IsComplex());
-            foreach (var property in properties)
-            {
-                var value = property.GetValue(exception);
-                if (value == null) { continue; }
-                Decorator.Enclose(writer).WriteObject(value, value.GetType(), o => o.Settings.RootName = new XmlQualifiedEntity(property.Name));
-            }
-
-            WriteInnerExceptions(writer, exception, includeStackTrace);
-        }
-
-        private static void WriteInnerExceptions(XmlWriter writer, Exception exception, bool includeStackTrace)
-        {
-            var innerExceptions = new List<Exception>();
-            if (exception is AggregateException aggregated)
-            {
-                innerExceptions.AddRange(aggregated.Flatten().InnerExceptions);
-            }
-            else
-            {
-                if (exception.InnerException != null) { innerExceptions.Add(exception.InnerException); }    
-            }
-            if (innerExceptions.Count > 0)
-            {
-                var endElementsToWrite = 0;
-                foreach (var inner in innerExceptions)
-                {
-                    var exceptionType = inner.GetType();
-                    writer.WriteStartElement(Decorator.Enclose(exceptionType.Name).SanitizeXmlElementName());
-                    if (exceptionType.Namespace != null) { writer.WriteAttributeString("namespace", exceptionType.Namespace); }
-                    WriteExceptionCore(writer, inner, includeStackTrace);
-                    endElementsToWrite++;
-                }
-                for (var i = 0; i < endElementsToWrite; i++) { writer.WriteEndElement(); }
-            }
         }
     }
 }
